// frontend/src/enums/JobStatus.ts
export enum JobStatus {
    NOT_STARTED = "NOT_STARTED",
    VISITING_FINN = "VISITING_FINN",
    VISITING_MEGLER = "VISITING_MEGLER",
    DOWNLOADING_PROSPECT = "DOWNLOADING_PROSPECT",
    CREATING_EMBEDDINGS = "CREATING_EMBEDDINGS",
    LLM_IN_PROGRESS = "LLM_IN_PROGRESS",
    FAILED = "FAILED",
    COMPLETED = "COMPLETED",
}